refactor: ResponseHandler 분리
